refactor: reorganize packages in infra-module
