+++ infrastructure/src/main/java/hanghae/infrastructure/adapter/JpaShowtimeRepositoryAdapter.java
@@ -1,14 +1,14 @@
 package hanghae.infrastructure.adapter;
 
+import hanghae.domain.domain.Movie;
 import hanghae.domain.domain.Showtime;
-import hanghae.domain.entity.ShowtimeEntity;
 import hanghae.domain.port.ShowtimeRepository;
+import hanghae.infrastructure.domain.entity.ShowtimeEntity;
 import hanghae.infrastructure.repository.JpaShowtimeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
@@ -17,14 +17,14 @@ public class JpaShowtimeRepositoryAdapter implements ShowtimeRepository {
     private final JpaShowtimeRepository jpaShowtimeRepository;
 
     @Override
-    public List<Showtime> findShowtimeByMovieId(Long movieId) {
-        return jpaShowtimeRepository.findShowtimeEntitiesByMovie_MovieId(movieId)
+    public List<Showtime> findShowtimeByMovie(Movie movie) {
+        return jpaShowtimeRepository.findShowtimeEntitiesByMovie_MovieId(movie.getMovieId())
                 .stream()
-                .map(this::toShowtimeFrom)
+                .map(showtimeEntity -> toShowtimeDomain(showtimeEntity, movie))
                 .toList();
     }
 
-    private Showtime toShowtimeFrom(ShowtimeEntity showtimeEntity) {
-        return Showtime.from(showtimeEntity);
+    private Showtime toShowtimeDomain(ShowtimeEntity showtimeEntity, Movie movie) {
+        return ShowtimeEntity.showtimeOf(showtimeEntity, movie);
     }
 }
