+++ infrastructure/src/main/java/hanghae/infrastructure/repository/JpaMovieRepository.java
@@ -1,6 +1,6 @@
 package hanghae.infrastructure.repository;
 
-import hanghae.domain.entity.MovieEntity;
+import hanghae.infrastructure.domain.entity.MovieEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
